Support dragging pieces around within the pool.
