+++ src/components/Board.js
@@ -81,3 +81,44 @@ export default function Board({
     </div>
   );
 }
+
+export function dragDestinationOnBoard(gameState, pointer) {
+  let boardRect = document.getElementById("board").getBoundingClientRect();
+  if (
+    gameState.dragState.destination.where === "board" ||
+    (boardRect.left <= pointer.x &&
+      pointer.x <= boardRect.right &&
+      boardRect.top <= pointer.y &&
+      pointer.y <= boardRect.bottom)
+  ) {
+    const draggedPieceIDs = gameState.dragState.pieceIDs;
+    const draggedPieces = gameState.pieces.filter((piece) =>
+      draggedPieceIDs.includes(piece.id)
+    );
+
+    const groupHeight = Math.max(
+      ...draggedPieces.map((piece) => piece.groupTop + piece.letters.length)
+    );
+    const groupWidth = Math.max(
+      ...draggedPieces.map((piece) => piece.groupLeft + piece.letters[0].length)
+    );
+    const maxTop = gameState.gridSize - groupHeight;
+    const maxLeft = gameState.gridSize - groupWidth;
+
+    const squareWidth = (boardRect.width - 1) / gameState.gridSize;
+    const squareHeight = (boardRect.height - 1) / gameState.gridSize;
+    const pointerOffset = gameState.dragState.pointerOffset;
+    const unclampedLeft = Math.round(
+      (pointer.x - pointerOffset.x - boardRect.left) / squareWidth
+    );
+    const unclampedTop = Math.round(
+      (pointer.y - pointerOffset.y - boardRect.top) / squareHeight
+    );
+    const left = Math.max(0, Math.min(maxLeft, unclampedLeft));
+    const top = Math.max(0, Math.min(maxTop, unclampedTop));
+
+    return { where: "board", top, left };
+  }
+
+  return undefined;
+}
