+++ src/components/DragGroup.js
@@ -1,5 +1,7 @@
 import Piece from "./Piece";
 import React from "react";
+import { dragDestinationOnBoard } from "./Board";
+import { dragDestinationInPool } from "./Pool";
 
 // This component is mounted each time a drag starts.
 export default function DragGroup({ dispatchGameState, gameState }) {
@@ -117,62 +119,10 @@ export default function DragGroup({ dispatchGameState, gameState }) {
 }
 
 function dragDestination(gameState, pointer) {
+  let destination = undefined;
   if (!gameState.dragState.isShifting) {
-    let poolElement =
-      document.getElementById("pool") || document.getElementById("result");
-    let poolRect = poolElement.getBoundingClientRect();
-    if (
-      poolRect.left <= pointer.x &&
-      pointer.x <= poolRect.right &&
-      poolRect.top <= pointer.y &&
-      pointer.y <= poolRect.bottom
-    ) {
-      if (gameState.dragState.destination.where === "pool") {
-        return gameState.dragState.destination;
-      }
-      let poolPieces = gameState.pieces.filter(
-        (piece) => piece.poolIndex >= 0
-      );
-      return { where: "pool", index: poolPieces.length };
-    }
+    destination ??= dragDestinationInPool(pointer);
   }
-
-  let boardRect = document.getElementById("board").getBoundingClientRect();
-  if (
-    gameState.dragState.destination.where === "board" ||
-    (boardRect.left <= pointer.x &&
-      pointer.x <= boardRect.right &&
-      boardRect.top <= pointer.y &&
-      pointer.y <= boardRect.bottom)
-  ) {
-    const draggedPieceIDs = gameState.dragState.pieceIDs;
-    const draggedPieces = gameState.pieces.filter((piece) =>
-      draggedPieceIDs.includes(piece.id)
-    );
-
-    const groupHeight = Math.max(
-      ...draggedPieces.map((piece) => piece.groupTop + piece.letters.length)
-    );
-    const groupWidth = Math.max(
-      ...draggedPieces.map((piece) => piece.groupLeft + piece.letters[0].length)
-    );
-    const maxTop = gameState.gridSize - groupHeight;
-    const maxLeft = gameState.gridSize - groupWidth;
-
-    const squareWidth = (boardRect.width - 1) / gameState.gridSize;
-    const squareHeight = (boardRect.height - 1) / gameState.gridSize;
-    const pointerOffset = gameState.dragState.pointerOffset;
-    const unclampedLeft = Math.round(
-      (pointer.x - pointerOffset.x - boardRect.left) / squareWidth
-    );
-    const unclampedTop = Math.round(
-      (pointer.y - pointerOffset.y - boardRect.top) / squareHeight
-    );
-    const left = Math.max(0, Math.min(maxLeft, unclampedLeft));
-    const top = Math.max(0, Math.min(maxTop, unclampedTop));
-
-    return { where: "board", top, left };
-  }
-
-  return undefined;
+  destination ??= dragDestinationOnBoard(gameState, pointer);
+  return destination;
 }
